change integer to big interger in create category_product table

+++ database/migrations/2019_07_16_184450_create_category_product_table.php
@@ -16,11 +16,11 @@ class CreateCategoryProductTable extends Migration
         Schema::create('category_product', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('product_id')->unsigned()->nullable();
+            $table->bigInteger('product_id')->unsigned()->nullable();
             $table->foreign('product_id')->references('id')
                 ->on('products')->onDelete('cascade');
 
-            $table->integer('category_id')->unsigned()->nullable();
+            $table->bigInteger('category_id')->unsigned()->nullable();
             $table->foreign('category_id')->references('id')
                 ->on('categories')->onDelete('cascade');
